Move Updater structs to shared and use it for postgres
